move NotYetImplementedException to API package to be reusable in plug-ins
