+++ cobigen/cobigen-core-parent/cobigen-core/src/main/java/com/capgemini/cobigen/impl/config/upgrade/ContextConfigurationUpgrader.java
@@ -6,9 +6,9 @@ import java.math.BigDecimal;
 import org.dozer.DozerBeanMapper;
 
 import com.capgemini.cobigen.api.constants.ConfigurationConstants;
+import com.capgemini.cobigen.api.exception.NotYetSupportedException;
 import com.capgemini.cobigen.impl.config.ContextConfiguration;
 import com.capgemini.cobigen.impl.config.constant.ContextConfigurationVersion;
-import com.capgemini.cobigen.impl.exceptions.NotYetSupportedException;
 
 /**
  * This class encompasses all logic for legacy context configuration detection and upgrading these to the
