+++ cobigen/cobigen-core-parent/cobigen-core/src/main/java/com/capgemini/cobigen/impl/config/versioning/VersionValidator.java
@@ -7,9 +7,9 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.capgemini.cobigen.api.exception.InvalidConfigurationException;
+import com.capgemini.cobigen.api.exception.NotYetSupportedException;
 import com.capgemini.cobigen.impl.config.constant.ContextConfigurationVersion;
 import com.capgemini.cobigen.impl.config.constant.TemplatesConfigurationVersion;
-import com.capgemini.cobigen.impl.exceptions.NotYetSupportedException;
 
 /**
  * The version validator checks the compatibility of CobiGen and its configuration files
